+++ migrations/20200412233226-create-orders.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.REAL
       },
       status: {
         type: Sequelize.STRING
